Upgrade ruby 3.0.0

+++ db/schema.rb
@@ -25,7 +25,7 @@ ActiveRecord::Schema.define(version: 2021_03_18_211620) do
     t.string "name", null: false
     t.text "body"
     t.string "record_type", null: false
-    t.integer "record_id", null: false
+    t.bigint "record_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["record_type", "record_id", "name"], name: "index_action_text_rich_texts_uniqueness", unique: true
